kstore user id preservation added

+++ composeApp/src/commonMain/kotlin/features/tasks/main/MainTasksContainer.kt
@@ -10,12 +10,14 @@ import pro.respawn.flowmvi.plugins.recover
 import pro.respawn.flowmvi.plugins.reduce
 import ru.kpfu.itis.common.mapper.ErrorMapper
 import ru.kpfu.itis.features.task.data.repository.TaskRepository
+import ru.kpfu.itis.features.task.data.store.UserStore
 import ru.kpfu.itis.features.task.domain.model.TaskModel
 
 class MainTasksContainer(
     private val errorMapper: ErrorMapper,
     private val repository: TaskRepository,
     private val configurationFactory: DefaultConfigurationFactory,
+    private val userStore: UserStore,
 ) : Container<TasksState<List<TaskModel>>, MainTasksIntent, MainTasksAction> {
 
     override val store: Store<TasksState<List<TaskModel>>, MainTasksIntent, MainTasksAction> =
@@ -24,43 +26,51 @@ class MainTasksContainer(
             configure(configurationFactory, "Tasks")
 
             recover { exception ->
-                updateState {
-                    TasksState.Error(errorMapper.map(exception = exception))
-                }
+                updateState { TasksState.Error(errorMapper.map(exception = exception)) }
                 null
             }
 
             reduce { intent ->
-                when (intent) {
-                    is MainTasksIntent.LoadActiveTasks -> {
-                        updateState { TasksState.Loading }
-                        val tasks = repository.getActiveTasks(2)
-                        updateState { TasksState.Success(tasks) }
-                    }
+                val userId = userStore.getUserId()
+                if (userId == null) {
+                    action(MainTasksAction.SignOut)
+                    intent(MainTasksIntent.ClearCache)
+                } else {
+                    when (intent) {
+                        is MainTasksIntent.LoadActiveTasks -> {
+                            updateState { TasksState.Loading }
+                            val tasks = repository.getActiveTasks(userId)
+                            updateState { TasksState.Success(tasks) }
+                        }
 
-                    is MainTasksIntent.EditTask -> {
-                        action(MainTasksAction.OpenTaskBottomSheet(intent.taskId))
-                    }
+                        is MainTasksIntent.EditTask -> {
+                            action(MainTasksAction.OpenTaskBottomSheet(intent.taskId))
+                        }
 
-                    is MainTasksIntent.DeleteTask -> {
-                        updateState { TasksState.Loading }
-                        repository.deleteTask(intent.taskId)
-                        val tasks = repository.getActiveTasks(2)
-                        updateState { TasksState.Success(tasks) }
-                    }
+                        is MainTasksIntent.DeleteTask -> {
+                            updateState { TasksState.Loading }
+                            repository.deleteTask(intent.taskId)
+                            val tasks = repository.getActiveTasks(userId)
+                            updateState { TasksState.Success(tasks) }
+                        }
 
-                    is MainTasksIntent.LoadAllTasks -> {
-                        updateState { TasksState.Loading }
-                        val tasks = repository.getActiveTasks(2)
-                        updateState { TasksState.Success(tasks) }
-                    }
+                        is MainTasksIntent.LoadAllTasks -> {
+                            updateState { TasksState.Loading }
+                            val tasks = repository.getActiveTasks(userId)
+                            updateState { TasksState.Success(tasks) }
+                        }
 
-                    MainTasksIntent.FloatingButtonClicked -> {
-                        action(MainTasksAction.OpenTaskBottomSheet())
-                    }
+                        MainTasksIntent.FloatingButtonClicked -> {
+                            action(MainTasksAction.OpenTaskBottomSheet())
+                        }
 
-                    is MainTasksIntent.ErrorOccurred -> {
-                        action(MainTasksAction.OpenErrorScreen(intent.errorModel, 2))
+                        is MainTasksIntent.ErrorOccurred -> {
+                            action(MainTasksAction.OpenErrorScreen(intent.errorModel, userId))
+                        }
+
+                        is MainTasksIntent.ClearCache -> {
+                            repository.clearTasks()
+                        }
                     }
                 }
             }
